pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "A personal AI for Tanav."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.5.2",
    "duckduckgo-search>=8.1.1",
    "langchain-community>=0.3.27",
    "pyaudio>=0.2.14",
    "pydantic-ai>=0.2.16",
    "python-dotenv>=1.1.0",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.3",
    "typer>=0.12.5",
    "rich>=13.7.1",
    "pywhatkit>=5.4",
    "chainlit>=1.2.0",
]

[project.scripts]
friday = "cli.__main__:main"
